translation-server/requirements.txt
# Core dependencies (必須)
flask==3.0.0
flask-cors==4.0.0
google-generativeai>=0.3.0
yt-dlp>=2023.12.30
cryptography>=41.0.0
requests>=2.31.0

# GUI dependencies (GUI機能用)
# Note: tkinterはPython標準ライブラリですが、一部環境では別途インストール必要
# Ubuntu/Debian: sudo apt-get install python3-tk
# CentOS/RHEL: sudo yum install tkinter
pystray>=0.19.0
pillow>=10.0.0

# Windows固有機能用 - 自動起動、多重起動防止
pywin32>=306; sys_platform == "win32"

# 開発・ビルド用
pyinstaller>=6.0.0